Organize the problem in more logical way
Previously, there was no requested data amount for UEs and it was
assumed to repeat in every period. Right now, the current problem
suggests that each UE has some amount of data to be requested, and
the main goal is to optimize the scheduling of these UEs, so that
all UEs have been placed in a good way energy-wise and latency-wise.